+++ ESFramework/Foundation/NSString+ESExtension.h
@@ -18,6 +18,11 @@ NS_ASSUME_NONNULL_BEGIN
 + (NSString *)UUIDString;
 
 /**
+ * Generates a random string that contains a-zA-Z0-9.
+ */
++ (nullable NSString *)randomStringWithLength:(NSUInteger)length;
+
+/**
  * Returns a NSNumber object created by parsing the string.
  */
 - (nullable NSNumber *)numberValue;