+++ ESFramework/Foundation/NSString+ESExtension.m
@@ -62,6 +62,24 @@
     return [NSUUID UUID].UUIDString;
 }
 
++ (nullable NSString *)randomStringWithLength:(NSUInteger)length
+{
+    static char charset[] = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+    static uint32_t charsetLen = (uint32_t)(sizeof(charset) - 1);
+
+    char *str = malloc(length + 1);
+    if (!str) {
+        return nil;
+    }
+
+    for (NSUInteger i = 0; i < length; i++) {
+        str[i] = charset[arc4random_uniform(charsetLen)];
+    }
+    str[length] = '\0';
+    
+    return [NSString stringWithUTF8String:str];
+}
+
 - (NSNumber *)numberValue
 {
     return [NSNumber numberWithString:self];